% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.shiny.league-package.R
\docType{package}
\name{nba.shiny.league-package}
\alias{nba.shiny.league}
\alias{nba.shiny.league-package}
\title{nba.shiny.league: NBA Shiny League}
\description{
A Shiny application for NBA Fantasy
}
\author{
\strong{Maintainer}: Oliver Eaton \email{eat_fred@protonmail.com}

Authors:
\itemize{
  \item Oliver Eaton \email{eat_fred@protonmail.com}
}

}
